Editor: refactor undo actions into new file
