+++ src/game/editor/editor.h
@@ -13,6 +13,7 @@
 #include <game/client/ui.h>
 #include <game/mapitems.h>
 #include <game/mapitems_ex.h>
+#include <game/editor/editor_actions.h>
 
 #include <engine/editor.h>
 #include <engine/graphics.h>
@@ -526,6 +527,19 @@ struct RECTi
 	int w, h;
 };
 
+struct CLayerTiles_SCommonPropState
+{
+	enum
+	{
+		MODIFIED_SIZE = 1 << 0,
+		MODIFIED_COLOR = 1 << 1,
+	};
+	int m_Modified = 0;
+	int m_Width = -1;
+	int m_Height = -1;
+	int m_Color = 0;
+};
+
 class CLayerTiles : public CLayer
 {
 protected:
@@ -613,18 +627,7 @@ public:
 	virtual void ShowInfo();
 	int RenderProperties(CUIRect *pToolbox) override;
 
-	struct SCommonPropState
-	{
-		enum
-		{
-			MODIFIED_SIZE = 1 << 0,
-			MODIFIED_COLOR = 1 << 1,
-		};
-		int m_Modified = 0;
-		int m_Width = -1;
-		int m_Height = -1;
-		int m_Color = 0;
-	};
+	using SCommonPropState = CLayerTiles_SCommonPropState;
 	static int RenderCommonProperties(SCommonPropState &State, CEditor *pEditor, CUIRect *pToolbox, std::vector<CLayerTiles *> &vpLayers);
 
 	void ModifyImageIndex(INDEX_MODIFY_FUNC pfnFunc) override;
@@ -704,491 +707,6 @@ public:
 	int RenderProperties(CUIRect *pToolbox) override;
 };
 
-class IEditorAction
-{
-public:
-	IEditorAction()
-	{
-		m_pEditor = nullptr;
-	}
-
-	virtual bool Undo() = 0;
-	virtual bool Redo() = 0;
-
-	virtual void Print()
-	{
-	}
-
-	virtual const char *Name() const = 0;
-
-	CEditor *m_pEditor;
-};
-
-template<typename T>
-class CEditorAction : public IEditorAction
-{
-public:
-	enum class EType
-	{
-		CHANGE_COLOR_TILE,
-		CHANGE_COLOR_QUAD,
-
-		EDIT_MULTIPLE_LAYERS,
-
-		ADD_LAYER,
-		DELETE_LAYER,
-
-		ADD_GROUP,
-		DELETE_GROUP,
-
-		EDIT_LAYER_GROUP,
-		EDIT_LAYER_ORDER,
-		EDIT_LAYER_WIDTH,
-		EDIT_LAYER_HEIGHT,
-		EDIT_LAYER_IMAGE,
-		EDIT_LAYER_COLOR_ENV,
-		EDIT_LAYER_COLOR_TO,
-
-		ADD_QUAD,
-		DELETE_QUAD,
-		EDIT_QUAD_POSITION,
-		EDIT_QUAD_CENTER,
-		EDIT_QUAD_VERTEX,
-
-		//SET_TILE,
-		FILL_SELECTION,
-		BRUSH_DRAW,
-
-		EDIT_GROUP_ORDER,
-		EDIT_GROUP_POS_X,
-		EDIT_GROUP_POS_Y,
-		EDIT_GROUP_PARA_X,
-		EDIT_GROUP_PARA_Y,
-		EDIT_GROUP_CLIP_X,
-		EDIT_GROUP_CLIP_Y,
-		EDIT_GROUP_CLIP_W,
-		EDIT_GROUP_CLIP_H,
-
-		ADD_ENVELOPPE,
-		EDIT_ENV_NAME,
-		EDIT_ENV_POINT_VALUE,
-		EDIT_ENV_POINT_TIME,
-		EDIT_ENV_SYNC_CHECKBOX,
-		DELETE_ENVELOPPE,
-
-		ADD_COMMAND,
-		EDIT_COMMAND,
-		DELETE_COMMAND,
-		MOVE_COMMAND_UP,
-		MOVE_COMMAND_DOWN,
-
-		//ADD_IMAGE,
-		//ADD_SOUND,
-
-		CHANGE_ORIENTATION_VALUE
-	};
-
-public:
-	CEditorAction(EType Type, void *pObject, const T &From, const T &To) :
-		IEditorAction()
-	{
-		m_ValueFrom = From;
-		m_ValueTo = To;
-		m_pObject = pObject;
-		m_Type = Type;
-	}
-
-	virtual ~CEditorAction()
-	{
-	}
-
-	const char *Name() const override
-	{
-		switch(m_Type)
-		{
-		case EType::CHANGE_COLOR_TILE: return "CHANGE_COLOR_TILE";
-		case EType::CHANGE_COLOR_QUAD: return "CHANGE_COLOR_QUAD";
-		case EType::EDIT_MULTIPLE_LAYERS: return "EDIT_MULTIPLE_LAYERS";
-		case EType::ADD_LAYER: return "ADD_LAYER";
-		case EType::DELETE_LAYER: return "DELETE_LAYER";
-		case EType::ADD_GROUP: return "ADD_GROUP";
-		case EType::DELETE_GROUP: return "DELETE_GROUP";
-		case EType::EDIT_LAYER_GROUP: return "EDIT_LAYER_GROUP";
-		case EType::EDIT_LAYER_ORDER: return "EDIT_LAYER_ORDER";
-		case EType::EDIT_LAYER_WIDTH: return "EDIT_LAYER_WIDTH";
-		case EType::EDIT_LAYER_HEIGHT: return "EDIT_LAYER_HEIGHT";
-		case EType::EDIT_LAYER_IMAGE: return "EDIT_LAYER_IMAGE";
-		case EType::EDIT_LAYER_COLOR_ENV: return "EDIT_LAYER_COLOR_ENV";
-		case EType::EDIT_LAYER_COLOR_TO: return "EDIT_LAYER_COLOR_TO";
-		case EType::EDIT_QUAD_POSITION: return "EDIT_QUAD_POSITION";
-		case EType::EDIT_QUAD_CENTER: return "EDIT_QUAD_CENTER";
-		case EType::FILL_SELECTION: return "FILL_SELECTION";
-		case EType::BRUSH_DRAW: return "BRUSH_DRAW";
-		case EType::EDIT_GROUP_ORDER: return "EDIT_GROUP_ORDER";
-		case EType::EDIT_GROUP_POS_X: return "EDIT_GROUP_POS_X";
-		case EType::EDIT_GROUP_POS_Y: return "EDIT_GROUP_POS_Y";
-		case EType::EDIT_GROUP_PARA_X: return "EDIT_GROUP_PARA_X";
-		case EType::EDIT_GROUP_PARA_Y: return "EDIT_GROUP_PARA_Y";
-		case EType::EDIT_GROUP_CLIP_X: return "EDIT_GROUP_CLIP_X";
-		case EType::EDIT_GROUP_CLIP_Y: return "EDIT_GROUP_CLIP_Y";
-		case EType::EDIT_GROUP_CLIP_W: return "EDIT_GROUP_CLIP_W";
-		case EType::EDIT_GROUP_CLIP_H: return "EDIT_GROUP_CLIP_H";
-		case EType::ADD_ENVELOPPE: return "ADD_ENVELOPPE";
-		case EType::EDIT_ENV_NAME: return "EDIT_ENV_NAME";
-		case EType::EDIT_ENV_POINT_VALUE: return "EDIT_ENV_POINT_VALUE";
-		case EType::EDIT_ENV_POINT_TIME: return "EDIT_ENV_POINT_TIME";
-		case EType::EDIT_ENV_SYNC_CHECKBOX: return "EDIT_ENV_SYNC_CHECKBOX";
-		case EType::DELETE_ENVELOPPE: return "DELETE_ENVELOPPE";
-		case EType::ADD_COMMAND: return "ADD_COMMAND";
-		case EType::EDIT_COMMAND: return "EDIT_COMMAND";
-		case EType::DELETE_COMMAND: return "DELETE_COMMAND";
-		case EType::MOVE_COMMAND_UP: return "MOVE_COMMAND_UP";
-		case EType::MOVE_COMMAND_DOWN: return "MOVE_COMMAND_DOWN";
-		//case EType::ADD_IMAGE: return "ADD_IMAGE";
-		//case EType::ADD_SOUND: return "ADD_SOUND";
-		case EType::CHANGE_ORIENTATION_VALUE: return "CHANGE_ORIENTATION_VALUE";
-		case EType::ADD_QUAD: return "ADD_QUAD";
-		case EType::DELETE_QUAD: return "DELETE_QUAD";
-		case EType::EDIT_QUAD_VERTEX: return "EDIT_QUAD_VERTEX";
-		default: return "UNKNOWN";
-		}
-	}
-
-	EType m_Type;
-
-protected:
-	T m_ValueFrom;
-	T m_ValueTo;
-	void *m_pObject;
-};
-
-template<typename T>
-class CEditorLayerAction : public CEditorAction<T>
-{
-public:
-	CEditorLayerAction(CEditorAction::EType Type, void *pObject, int GroupIndex, int LayerIndex, const T &From, const T &To) :
-		CEditorAction<T>(Type, pObject, From, To)
-	{
-		m_GroupIndex = GroupIndex;
-		m_LayerIndex = LayerIndex;
-	}
-
-	template<typename K>
-	K *GetLayer();
-
-protected:
-	int m_GroupIndex;
-	int m_LayerIndex;
-};
-
-template<typename T>
-template<typename K>
-K* CEditorLayerAction<T>::GetLayer()
-{
-	return (K *)(m_pEditor->m_Map.m_vpGroups[m_GroupIndex]->m_vpLayers[m_LayerIndex]);
-}
-
-class CEditorChangeColorTileAction : public CEditorLayerAction<CColor>
-{
-public:
-	CEditorChangeColorTileAction(int GroupIndex, int LayerIndex, const CColor &OldColor, const CColor &NewColor) :
-		CEditorLayerAction(CEditorAction::EType::CHANGE_COLOR_TILE, nullptr, GroupIndex, LayerIndex, OldColor, NewColor)
-	{
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-	void Print() override
-	{
-		dbg_msg("editor", "Editor action: Change tile color, prev=%d %d %d %d, new=%d %d %d %d", m_ValueFrom.r, m_ValueFrom.g, m_ValueFrom.b, m_ValueFrom.a, m_ValueTo.r, m_ValueTo.g, m_ValueTo.b, m_ValueTo.a);
-	}
-};
-
-class CEditorChangeColorQuadAction : public CEditorAction<CColor>
-{
-public:
-	CEditorChangeColorQuadAction(void *pObject, int Index, const CColor &OldColor, const CColor &NewColor) :
-		CEditorAction(CEditorAction::EType::CHANGE_COLOR_QUAD, pObject, OldColor, NewColor)
-	{
-		m_Index = Index;
-	}
-
-	bool Undo() override
-	{
-		CQuad *pQuad = (CQuad *)m_pObject;
-		pQuad->m_aColors[m_Index] = m_ValueFrom;
-		return true;
-	}
-
-	bool Redo() override
-	{
-		CQuad *pQuad = (CQuad *)m_pObject;
-		pQuad->m_aColors[m_Index] = m_ValueTo;
-		return true;
-	}
-
-	void Print() override
-	{
-		dbg_msg("editor", "Editor action: Change quad color, vertex %d, prev=%d %d %d %d, new=%d %d %d %d", m_Index, m_ValueFrom.r, m_ValueFrom.g, m_ValueFrom.b, m_ValueFrom.a, m_ValueTo.r, m_ValueTo.g, m_ValueTo.b, m_ValueTo.a);
-	}
-
-private:
-	int m_Index;
-};
-
-class CEditorAddLayerAction : public CEditorAction<CLayer *>
-{
-public:
-	CEditorAddLayerAction(int GroupIndex, int LayerIndex, std::function<void()> fnAddLayer) :
-		CEditorAction(CEditorAction::EType::ADD_LAYER, nullptr, nullptr, nullptr)
-	{
-		m_GroupIndex = GroupIndex;
-		m_LayerIndex = LayerIndex;
-		m_fnAddLayer = fnAddLayer;
-	}
-
-	~CEditorAddLayerAction()
-	{
-	}
-
-	bool Undo() override;
-
-	bool Redo() override
-	{
-		// call the add layer function, needed for special layers
-		m_fnAddLayer();
-		return true;
-	}
-
-	void Print() override
-	{
-		//dbg_msg("editor", "Editor action: Add layer, type=%d name=%s", m_ValueTo->m_Type, m_ValueTo->m_aName);
-	}
-
-private:
-	int m_LayerIndex;
-	int m_GroupIndex;
-	std::function<void()> m_fnAddLayer;
-};
-
-class CEditorDeleteLayerAction : public CEditorAction<CLayer *>
-{
-public:
-	CEditorDeleteLayerAction(CEditor *pEditor, int GroupIndex, int LayerIndex);
-
-	~CEditorDeleteLayerAction()
-	{
-		delete m_ValueFrom;
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-	void Print() override
-	{
-		dbg_msg("editor", "Editor action: Delete layer, type=%d name=%s", m_ValueFrom->m_Type, m_ValueFrom->m_aName);
-	}
-
-private:
-	int m_LayerIndex;
-	int m_GroupIndex;
-};
-
-class CEditorEditMultipleLayersAction : public CEditorAction<std::vector<CLayerTiles::SCommonPropState>>
-{
-public:
-	CEditorEditMultipleLayersAction(void *pObject, std::vector<CLayerTiles::SCommonPropState> vOriginals, CLayerTiles::SCommonPropState State, int GroupIndex, std::vector<int> vLayers) :
-		CEditorAction(CEditorAction::EType::EDIT_MULTIPLE_LAYERS, pObject, vOriginals, {State})
-	{
-		m_GroupIndex = GroupIndex;
-		m_vLayers = vLayers;
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-	void Print() override
-	{
-		dbg_msg("editor", "Editor action: Edit multiple layers");
-	}
-
-private:
-	std::vector<int> m_vLayers;
-	int m_GroupIndex;
-};
-
-class CEditorAddGroupAction : public CEditorAction<int>
-{
-public:
-	CEditorAddGroupAction(CEditorMap *pObject, int GroupIndex) :
-		CEditorAction(CEditorAction::EType::ADD_GROUP, pObject, -1, GroupIndex)
-	{
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-};
-
-class CEditorDeleteGroupAction : public CEditorAction<CLayerGroup *>
-{
-public:
-	CEditorDeleteGroupAction(CEditorMap *pObject, int GroupIndex) :
-		CEditorAction(CEditorAction::EType::DELETE_GROUP, pObject, new CLayerGroup(*pObject->m_vpGroups[GroupIndex]), nullptr)
-	{
-		m_GroupIndex = GroupIndex;
-	}
-
-	~CEditorDeleteGroupAction()
-	{
-		delete m_ValueFrom;
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-private:
-	int m_GroupIndex;
-};
-
-class CEditorFillSelectionAction : public CEditorAction<CLayerGroup *>
-{
-public:
-	CEditorFillSelectionAction(void *pObject, CLayerGroup *Original, CLayerGroup *Brush, int GroupIndex, std::vector<int> vLayers, CUIRect Rect) :
-		CEditorAction(CEditorAction::EType::FILL_SELECTION, pObject, new CLayerGroup(*Original), new CLayerGroup(*Brush))
-	{
-		m_Rect = Rect;
-		m_vLayers = vLayers;
-		m_GroupIndex = GroupIndex;
-	}
-
-	~CEditorFillSelectionAction()
-	{
-		delete m_ValueFrom;
-		delete m_ValueTo;
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-	void Print() override;
-
-private:
-	CUIRect m_Rect;
-	int m_GroupIndex;
-	std::vector<int> m_vLayers;
-};
-
-class CEditorBrushDrawAction : public CEditorAction<CLayerGroup *>
-{
-public:
-	CEditorBrushDrawAction(void *pObject, CLayerGroup *Original, CLayerGroup *Brush, int GroupIndex, std::vector<int> vLayers) :
-		CEditorAction(CEditorAction::EType::BRUSH_DRAW, pObject, new CLayerGroup(*Original), new CLayerGroup(*Brush))
-	{
-		m_vLayers = vLayers;
-		m_GroupIndex = GroupIndex;
-	}
-
-	~CEditorBrushDrawAction()
-	{
-		delete m_ValueFrom;
-		delete m_ValueTo;
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-	void Print() override;
-
-private:
-	std::vector<int> m_vLayers;
-	int m_GroupIndex;
-
-	bool BrushDraw(CLayerGroup *Brush);
-};
-
-class CEditorCommandAction : public CEditorAction<std::string>
-{
-public:
-	CEditorCommandAction(CEditorAction::EType Action, int* pSelectedCommand, int CommandIndex, std::string Old, std::string New) :
-		CEditorAction(Action, nullptr, Old, New)
-	{
-		m_pSelectedCommand = pSelectedCommand;
-		m_CommandIndex = CommandIndex;
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-private:
-	int m_CommandIndex;
-	int *m_pSelectedCommand;
-};
-
-class CEditorAddQuadAction : public CEditorLayerAction<RECTi>
-{
-public:
-	CEditorAddQuadAction(int GroupIndex, int LayerIndex, RECTi Quad) :
-		CEditorLayerAction(CEditorAction::EType::ADD_QUAD, nullptr, GroupIndex, LayerIndex, {}, Quad)
-	{
-	}
-
-	bool Undo() override;
-	bool Redo() override;
-
-};
-
-class CEditorDeleteQuadsAction : public CEditorLayerAction<void*>
-{
-public:
-	CEditorDeleteQuadsAction(int GroupIndex, int LayerIndex, std::vector<int> vQuadIndexes, std::vector<CQuad> vQuads) :
-		CEditorLayerAction(CEditorAction::EType::DELETE_QUAD, nullptr, GroupIndex, LayerIndex, nullptr, nullptr)
-	{
-		m_vQuadIndexes = vQuadIndexes;
-		m_vQuads = vQuads;
-	}
-
-	bool Undo() override
-	{
-		CLayerQuads *pLayer = GetLayer<CLayerQuads>();
-
-		size_t k = 0;
-		
-		pLayer->m_vQuads.resize(pLayer->m_vQuads.size() + m_vQuads.size());
-
-		for(auto &Index : m_vQuadIndexes)
-		{
-			pLayer->m_vQuads.insert(pLayer->m_vQuads.begin() + Index, m_vQuads[k]);
-			k++;
-		}
-
-		return true;
-	}
-
-	bool Redo() override
-	{
-		CLayerQuads *pLayer = GetLayer<CLayerQuads>();
-		std::vector vSelectedQuads = m_vQuadIndexes;
-
-		for(int i = 0; i < (int)vSelectedQuads.size(); ++i)
-		{
-			pLayer->m_vQuads.erase(pLayer->m_vQuads.begin() + vSelectedQuads[i]);
-			for(int j = i + 1; j < (int)vSelectedQuads.size(); ++j)
-				if(vSelectedQuads[j] > vSelectedQuads[i])
-					vSelectedQuads[j]--;
-
-			vSelectedQuads.erase(vSelectedQuads.begin() + i);
-			i--;
-		}
-
-		return true;
-	}
-
-private:
-	std::vector<int> m_vQuadIndexes;
-	std::vector<CQuad> m_vQuads;
-};
-
 class CEditorHistory
 {
 public:
@@ -1202,7 +720,7 @@ public:
 		Clear();
 	}
 
-	void RecordUndoAction(IEditorAction *pAction, bool Clear = true);
+	void RecordUndoAction(class IEditorAction *pAction, bool Clear = true);
 	bool Undo();
 	bool Redo();
 
@@ -1216,7 +734,7 @@ public:
 	std::deque<IEditorAction *> m_vpRedoActions;
 
 private:
-	void RecordRedoAction(IEditorAction *Action);
+	void RecordRedoAction(class IEditorAction *Action);
 
 };
 
@@ -1595,7 +1113,7 @@ public:
 	struct CLayerPopupContext
 	{
 		std::vector<CLayerTiles *> m_vpLayers;
-		CLayerTiles::SCommonPropState m_CommonPropState;
+		CLayerTiles_SCommonPropState m_CommonPropState;
 	};
 	static int PopupLayer(CEditor *pEditor, CUIRect View, void *pContext);
 	static int PopupQuad(CEditor *pEditor, CUIRect View, void *pContext);
